[Feature #161214122] added endpoint to retrieve single sale record
