+++ app/__init__.py
@@ -2,12 +2,11 @@ from flask import Blueprint
 from flask_restful import Api
 
 # local imports
-from app.api.v1.views import Home, Products
+from app.api.v1.sales_view import SingleRecord
 
 
 api_bp = Blueprint('api', __name__)
 api = Api(api_bp)
 
 # Routes
-api.add_resource(Home, '/home')
-api.add_resource(Products, '/products')
+api.add_resource(SingleRecord, '/sales/<int:saleID>')
